requirements.txt
Django==5.0.7
djangorestframework==3.15.2
django-cors-headers==4.4.0
psycopg[binary]==3.2.1
Pillow==10.4.0
